usb: dwc3: Avoid resume_work flush in pm_freeze/restore
dwc3_resume_work is dependent on sm_work which is a
freezable task. Therefore when control reaches
pm_freeze/restore, sm_work is frozen as part of freeze_processes()
called from system-wide suspend. So if any pending
work on sm_work is there and we try to flush, it will end up
in deadlock as freeze/restore cannot progress.

Fix this be avoiding flush work during pm_freeze/restore.

Change-Id: Icd19c6d9d53d5a804644c5584667d95929ecce6c
Signed-off-by: Udipto Goswami <ugoswami@codeaurora.org>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>